auth-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
pydantic==2.7.1
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.0.1
